requirements.txt
astroid==2.4.2
certifi==2020.4.5.2
chardet==3.0.4
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pkg-resources==0.0.0
pylint==2.5.3
python-dotenv==0.13.0
requests==2.23.0
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.9
wrapt==1.12.1
